use to yarn workspaces
The frontend and backend are two completely different beasts with two
completely different test harnesses, two completely different sets of
dependencies, and two completely different deployment targets, so it
makes sense that they should be treated as two completely separate
targets.

This uses yarn workspaces to partition backend and frontend as well as
adding a mocha test suite to the backend since we are not currently
able to use bigtest yet in node.