create a test server for easy scenario creation
We're going to need to sequence a lot of different events happening at
different times and a way to simulate these sequences happening in
tests.

This creates a `TestServer` resource that can be used to deliver
simulated webhooks at the http layer. The `server.deliver()` method
will actually create a webhook delivery, create the signature headers
and then transmit the hook over an http request.