introduce express
We're preaparing to add a GraphQL interface as a window into the state
of our release app, so we're going to need to be able to serve
multiple paths from the same http port. The logical move is to use
express insteod of node's `http` module.

This migrates to express using a channel per route to handle
requests. In other words, the route callback will send the request to
a special channel that is set up for that route. The handler for the
route can then subscribe to the channel.

The current github webhooks are then moved onto the `/github-webhook`
path.